backend/requirements.txt
fastapi
uvicorn
python-whois
requests
python-dotenv
httpx
google-genai
beautifulsoup4
